Lookup new column names for an accident dataset

This function retreieves the new column names corresponding to the original column names in a specified dataset. The column names are based on the configuration loaded by load_config(). This is useful when the original names are in Japanese (2-byte characters), which can be difficult to handle in programs. This function converts them to corresponding English names.

Usage

lookup_column_names(dataset_name, original_names)

Arguments

dataset_name

The name of the dataset.

original_names

A vector of original column names.

Value

A vector of new column names corresponding to the provided original names.

Examples

lookup_column_names("main_data", c("死者数", "天候"))
#> [1] "faitality_number" "weather"