Post-process data

These function perform post-processing operations on different datasets. post_process_main processes the main data (honhyo), separating it into accident and person data, and returning them as a list. post_process_sub and post_process_highway currently do not perform any modifications on sub data and highway data, respectively.

Usage

post_process_main(data)

post_process_sub(data)

post_process_highway(data)

Arguments

data

The data to be processed.

Value

A processed data.